[Feat]#175-HomeView Sync 수정
